requirements.txt
requests==2.23.0
#GRPC requirements
grpcio>=1.28.1
grpcio-tools>=1.28.1
protobuf>=3.11.3
six>=1.14.0
wheel>=0.29
psutil>=5.5.0
flask>=1.1.2
bintrees>=2.1.0
numpy
